+++ packages/core/src/infra/docker.rs
@@ -145,6 +145,45 @@ pub(crate) async fn pull_and_start_container(
     Ok(())
 }
 
+pub async fn get_container_logs(
+    docker: &Docker,
+    container_name: &str,
+    tail_lines: Option<usize>,
+) -> Result<Vec<String>> {
+    let tail = tail_lines
+        .map(|n| n.to_string())
+        .unwrap_or_else(|| "all".to_string());
+
+    let options = bollard::container::LogsOptions::<String> {
+        stdout: true,
+        stderr: true,
+        follow: false,
+        timestamps: true,
+        tail,
+        ..Default::default()
+    };
+
+    let logs = docker.logs(container_name, Some(options));
+    let mut log_strings = Vec::new();
+
+    let mut stream = logs;
+
+    while let Some(result) = stream.next().await {
+        match result {
+            Ok(output) => match output {
+                bollard::container::LogOutput::StdOut { message }
+                | bollard::container::LogOutput::StdErr { message } => {
+                    log_strings.push(String::from_utf8_lossy(&message).to_string());
+                }
+                _ => {}
+            },
+            Err(e) => return Err(e.into()),
+        }
+    }
+
+    Ok(log_strings)
+}
+
 fn create_binding_string(binding: &Binding) -> String {
     match &binding.options {
         Some(options) => format!("{}:{}:{}", binding.source, binding.destination, options),
